[#153] Introduce common.rs and replace posix shm and process local impl with type alias
